requirements.txt
streamlit
requests
pydantic
firecrawl-py
agno
google-generativeai
google-genai
